Add prototyping notebook into experimentation dir.
The prototyping is done on a Jupyter notebook. It makes use
of the Berkeley Enron database of emails. The aim is to create the
groundwork for the actual classifier and experiment with different
classifiers, parameters, feature selection methods, etc. Note that
much of the documentation for the experimentation process is given
within the notebook.